requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface-hub
tiktoken
#InstructorEmbedding
#sentence_transformers
langchain-community